+++ core/src/main/scala/samoyed/core/model/db/UserReleaseNotification.scala
@@ -4,9 +4,9 @@ import scalikejdbc._
 import java.time.OffsetDateTime
 
 case class UserReleaseNotification(
-    id: Int,
-    userId: Int,
-    artistAlbumId: Int,
+    id: Long,
+    userId: Long,
+    artistAlbumId: Long,
     createdAt: OffsetDateTime,
     updatedAt: OffsetDateTime,
     deletedAt: Option[OffsetDateTime],
